# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventarios de activos",
        "description": "APIs para el manejo de inventarios de activos",
        "version": "1.0.0"
    },
    "paths": {
...
        "/digital/api/actions/{id}": {
            "get": {
                "tags": [
                    "Actions"
                ],
                "summary": "Get an action",
                "description": "Get details of an action from the system",
                "operationId": "GetAction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the action",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No data found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Actions"
                ],
                "summary": "Delete an action",
                "description": "Delete an action from the system",
                "operationId": "DeleteAction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the action",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource deleted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No data found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/digital/api/assets": {
            "post": {
                "tags": [
                    "Assets"
                ],
                "summary": "Store a new asset",
                "description": "Store a new asset in the system",
                "operationId": "StoreAsset",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "code",
                                    "name",
                                    "type_id",
                                    "location_id",
                                    "department_id",
                                    "status_id",
                                    "classification_id",
                                    "criticality_id",
                                    "action_id",
                                    "created_by"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "type_id": {
                                        "type": "integer"
                                    },
                                    "location_id": {
                                        "type": "integer"
                                    },
                                    "department_id": {
                                        "type": "integer"
                                    },
                                    "status_id": {
                                        "type": "integer"
                                    },
                                    "classification_id": {
                                        "type": "integer"
                                    },
                                    "criticality_id": {
                                        "type": "integer"
                                    },
                                    "action_id": {
                                        "type": "integer"
                                    },
                                    "created_by": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/digital/api/assets/{id}": {
            "get": {
                "tags": [
                    "Assets"
                ],
                "summary": "Get an asset",
                "description": "Get details of an asset from the system",
                "operationId": "GetAsset",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the asset",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No data found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Assets"
                ],
                "summary": "Delete an asset",
                "description": "Delete an asset from the system",
                "operationId": "DeleteAsset",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the asset",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource deleted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No data found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/digital/api/classifications": {
            "post": {
                "tags": [
                    "Classifications"
                ],
                "summary": "Store a new classification",
                "description": "Store a new classification in the system",
                "operationId": "StoreClassification",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "classification_name"
                                ],
                                "properties": {
                                    "classification_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/digital/api/classifications/{id}": {
            "get": {
                "tags": [
                    "Classifications"
                ],
                "summary": "Get a classification",
                "description": "Get details of a classification from the system",
                "operationId": "GetClassification",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the classification",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No data found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Classifications"
                ],
                "summary": "Delete a classification",
                "description": "Delete a classification from the system",
                "operationId": "DeleteClassification",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the classification",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource deleted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No data found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}